requirements.txt
requests>=2.28
python-dotenv>=1.0
tenacity>=8.2.0
pytest>=7.0